+++ haystack/preview/__init__.py
@@ -1,0 +1,2 @@
+from canals import node
+from haystack.preview.pipeline import Pipeline, PipelineError, NoSuchStoreError, load_pipelines, save_pipelines
