+++ haystack/preview/rest_api/pipelines/default.json
@@ -1,0 +1,4 @@
+{
+    "pipelines": {},
+    "dependencies": ["haystack.preview"]
+}
