+++ haystack/preview/rest_api/routers/__init__.py
@@ -1,0 +1,3 @@
+from haystack.preview.rest_api.routers.about import router
+from haystack.preview.rest_api.routers.pipelines import router
+from haystack.preview.rest_api.routers.files import router
