--- haystack/v2/__init__.py
@@ -1,2 +1,0 @@
-from canals import node
-from haystack.v2.pipeline import Pipeline, PipelineError, NoSuchStoreError, load_pipelines, save_pipelines
