--- haystack/v2/rest_api/__init__.py
@@ -1,2 +1,0 @@
-from haystack.v2.rest_api.test.nodes.add_value import AddValue
-from haystack.v2.rest_api.test.nodes.double import Double
