--- haystack/v2/rest_api/routers/__init__.py
@@ -1,3 +1,0 @@
-from haystack.v2.rest_api.routers.about import router
-from haystack.v2.rest_api.routers.pipelines import router
-from haystack.v2.rest_api.routers.files import router
