--- haystack/v2/rest_api/test/conftest.py
@@ -1,11 +1,0 @@
-import pytest
-from fastapi.testclient import TestClient
-
-from haystack.v2.rest_api.app import get_app
-
-
-@pytest.fixture(autouse=True)
-def client():
-    app = get_app()
-    client = TestClient(app)
-    return client
