+++ test/preview/rest_api/nodes/__init__.py
@@ -1,0 +1,2 @@
+from haystack.preview.rest_api.test.nodes.add_value import AddValue
+from haystack.preview.rest_api.test.nodes.double import Double
